Move rh_cloud_utils methods to robottelo/utils/io
