+++ robottelo/utils/io/__init__.py
@@ -1,35 +1,8 @@
-"""Utility module for RH cloud inventory tests"""
+# Helper methods for tests requiring I/0
 import hashlib
 import json
 import tarfile
 from pathlib import Path
-
-
-def get_host_counts(tarobj):
-    """Returns hosts count from tar file.
-
-    Args:
-        tarobj: tar file to get host count from
-    """
-    metadata_counts = {}
-    slices_counts = {}
-    for file_ in tarobj.getmembers():
-        file_name = Path(file_.name).name
-        if not file_name.endswith('.json'):
-            continue
-        json_data = json.load(tarobj.extractfile(file_))
-        if file_name == 'metadata.json':
-            metadata_counts = {
-                f'{key}.json': value['number_hosts']
-                for key, value in json_data['report_slices'].items()
-            }
-        else:
-            slices_counts[file_name] = len(json_data['hosts'])
-
-    return {
-        'metadata_counts': metadata_counts,
-        'slices_counts': slices_counts,
-    }
 
 
 def get_local_file_data(path):
@@ -61,6 +34,33 @@ def get_local_file_data(path):
         'extractable': extractable,
         'json_files_parsable': json_files_parsable,
         **host_counts,
+    }
+
+
+def get_host_counts(tarobj):
+    """Returns hosts count from tar file.
+
+    Args:
+        tarobj: tar file to get host count from
+    """
+    metadata_counts = {}
+    slices_counts = {}
+    for file_ in tarobj.getmembers():
+        file_name = Path(file_.name).name
+        if not file_name.endswith('.json'):
+            continue
+        json_data = json.load(tarobj.extractfile(file_))
+        if file_name == 'metadata.json':
+            metadata_counts = {
+                f'{key}.json': value['number_hosts']
+                for key, value in json_data['report_slices'].items()
+            }
+        else:
+            slices_counts[file_name] = len(json_data['hosts'])
+
+    return {
+        'metadata_counts': metadata_counts,
+        'slices_counts': slices_counts,
     }
